added heatmap of damage
